Revert "Revert "change db to postgres""
This reverts commit da0ee202ff762a6357f884a996cb2d9103b3c212.

+++ jobmine/db/migrate/20171125214748_add_latitude_to_jobs.rb
@@ -1,5 +1,5 @@
 class AddLatitudeToJobs < ActiveRecord::Migration[5.1]
   def change
-    add_column :jobs, :latitude, :double
+    add_column :jobs, :latitude, :decimal
   end
 end
