+++ jobmine/db/migrate/20171125215007_add_longitube_to_jobs.rb
@@ -1,5 +1,5 @@
 class AddLongitubeToJobs < ActiveRecord::Migration[5.1]
   def change
-    add_column :jobs, :longitude, :double
+    add_column :jobs, :longitude, :decimal
   end
 end
